+++ src/ska_tmc_dishleafnode/commands/track_command.py
@@ -121,6 +121,13 @@ class Track(DishLNCommand):
             result_code, message = self.call_adapter_method(
                 "Dish Master", self.dish_master_adapter, "Track"
             )
+            self.component_manager.command_mapping.setdefault(
+                self.component_manager.command_id, {}
+            )["message_or_unique_id"] = message[0]
+            self.logger.debug(
+                "self.component_manager.command_mapping: %s",
+                self.component_manager.command_mapping,
+            )
 
         if result_code[0] == ResultCode.FAILED:
             return result_code[0], message[0]