+++ tests/unit/test_track_load_static_off_command.py
@@ -125,7 +125,6 @@ def test_configure_command_completed_with_correction_key_reset(
     task_callback.assert_against_call(
         call_kwargs={"status": TaskStatus.IN_PROGRESS}
     )
-
     task_callback.assert_against_call(
         call_kwargs={
             "status": TaskStatus.COMPLETED,
@@ -148,6 +147,8 @@ def test_configure_command_completed_with_correction_key_reset(
         time.sleep(1)
 
     assert "Command Completed" in message
+    cm.set_track_process_event()
+    cm.stop_track_table_process()
 
 
 def test_configure_command_completed_with_correction_key_update(
@@ -159,6 +160,7 @@ def test_configure_command_completed_with_correction_key_update(
 ):
     """Test configure command with correction key as UPDATE"""
 
+    cm.get_device().update_unresponsive(False, "")
     dish_device = DevFactory().get_device(DISH_MASTER_DEVICE)
     set_kvalue_command = SetKValue(cm, logger=logger)
     result_code, _ = set_kvalue_command.do(1)
@@ -222,6 +224,9 @@ def test_configure_command_completed_with_correction_key_update(
         time.sleep(1)
 
     assert "Command Completed" in message
+    time.sleep(0.2)
+    cm.set_track_process_event()
+    cm.stop_track_table_process()
 
 
 def test_correction_key_reset_partial_config(